+++ CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Do not try to generate body params for non-application/json request bodies
+  and also remove validation to prevent non-application/json bodies, this allows
+  a handler to handle the body in whichever way they desire, while still using
+  the open api spec to its full potential.
+
 ## [v0.0.27] - 2022-07-21
 
 ### Fixed